requirements.txt
mcp[cli]
openpyxl
python-dotenv
pydantic